refactor(core): extract debug_snapshot into parish-diagnostics crate
Move the debug-snapshot builders (debug_snapshot/) and the bug-report
orchestration (ipc/bug_report.rs) out of parish-core into a new
backend-agnostic leaf crate, parish-diagnostics. Behaviour-preserving:
parish-core re-exports both under their historical paths
(parish_core::debug_snapshot::* and parish_core::ipc::bug_report::*) so
every consumer (parish-server, parish-tauri, parish-harness, parish-mcp,
parish-engine) compiles with zero import changes.

Two reach-back couplings to parish-core are inverted via traits defined in
the new crate and implemented by parish-core, avoiding a dependency cycle:
- InferenceCategoryConfig (for GameConfig) — build_inference_categories.
- WorldSnapshotFields (for WorldSnapshot) — BugReportState::from_snapshots.
The shared weekday_name helper moves to parish_types::time.

parish-diagnostics is added to the architecture-fitness BACKEND_AGNOSTIC
list. Workspace grows to 22 crates / 13 leaf crates. Docs (AGENTS.md,
README.md, architecture.md, codebase-map.md) and the new crate README
updated. No new third-party dependency, so THIRD_PARTY_NOTICES is unchanged.
