+++ parish/Cargo.toml
@@ -18,6 +18,7 @@ members = [
     "crates/parish-npc-tool",
     "crates/parish-mod",
     "crates/parish-editor",
+    "crates/parish-diagnostics",
     "crates/parish-persistence",
     "crates/parish-input",
     "crates/parish-client",
@@ -51,6 +52,7 @@ parish-palette = { path = "crates/parish-palette" }
 parish-npc = { path = "crates/parish-npc" }
 parish-mod = { path = "crates/parish-mod" }
 parish-editor = { path = "crates/parish-editor" }
+parish-diagnostics = { path = "crates/parish-diagnostics" }
 parish-persistence = { path = "crates/parish-persistence" }
 parish-input = { path = "crates/parish-input" }
 parish-core = { path = "crates/parish-core" }
