--- parish/crates/parish-core/src/debug_snapshot/reexport.rs
@@ -1,1 +1,0 @@
-pub use crate::inference::InferenceLogEntry;
