+++ parish/crates/parish-diagnostics/Cargo.toml
@@ -1,39 +1,29 @@
 [package]
-name = "parish-core"
+name = "parish-diagnostics"
 version = "0.1.0"
 edition = "2024"
 license = "GPL-3.0-only"
 publish = false
-description = "Orchestration layer for the Parish engine: game session, IPC, mod loading, prompts"
+description = "Backend-agnostic diagnostics for Parish: debug-snapshot builders and bug-report orchestration"
 
 [dependencies]
 parish-types = { workspace = true }
 parish-config = { workspace = true }
 parish-inference = { workspace = true }
 parish-world = { workspace = true }
-parish-palette = { workspace = true }
 parish-npc = { workspace = true }
-parish-mod = { workspace = true }
-parish-editor = { workspace = true }
-parish-diagnostics = { workspace = true }
-parish-persistence = { workspace = true }
-parish-input = { workspace = true }
 tokio = { workspace = true }
-tokio-util = { workspace = true }
+# reqwest stays a direct dep: bug_report.rs makes its own GitHub REST calls
+# (issue create + screenshot upload). Only the `json` feature is needed.
 reqwest = { workspace = true, features = ["json"] }
 serde = { workspace = true }
 serde_json = { workspace = true }
-serde_yaml = { workspace = true }
-anyhow = { workspace = true }
+chrono = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
-chrono = { workspace = true }
-rand = { workspace = true }
-toml = { workspace = true }
-base64 = "0.22"
 uuid = { workspace = true }
+base64 = "0.22"
 
 [dev-dependencies]
 tempfile = { workspace = true }
 wiremock = { workspace = true }
-regex = { workspace = true }
