+++ parish/crates/parish-diagnostics/src/debug_snapshot/reexport.rs
@@ -1,0 +1,1 @@
+pub use parish_inference::InferenceLogEntry;
