+++ parish/testing/fixtures/play_extract-diagnostics.txt
@@ -1,0 +1,65 @@
+# Verification fixture for extract-diagnostics.
+#
+# Exercises the game through the paths that touch `debug_snapshot` and
+# `bug_report` functionality: status, clock, NPC introspection, movement
+# (which pushes DebugEvents), and the /debug commands that call the
+# debug-snapshot builders. Proves behavior parity — the output must be
+# structurally identical before and after moving those modules to
+# parish-diagnostics.
+#
+# NOTE: The snapshot/bug-report HTTP criteria (C7, C8) cannot be proven
+# by this headless fixture because the harness runs the engine in-process
+# without an HTTP layer. Those criteria are proven separately in
+# evidence.md via:
+#   - curl http://localhost:3030/api/debug-snapshot  (live server)
+#   - parish_file_bug MCP call with PARISH_BUG_REPORT_DRY_RUN=1
+#
+# Run:
+#   cargo run --manifest-path parish/Cargo.toml -p parish-engine \
+#     -- --script parish/testing/fixtures/play_extract-diagnostics.txt \
+#     > .proofs/extract-diagnostics/transcript.txt
+#
+# Expected signals:
+#   - "kind":"status" or equivalent (engine alive, C6 baseline)
+#   - game_time field (clock module loaded, C6 clock parity)
+#   - player_location field or location name (world graph loaded, C6)
+#   - /debug clock: game_time + season present (debug_snapshot builder, C6)
+#   - /debug npcs: tier or name field present (NPC debug builder, C6)
+#   - Movement response: player moved to new location (DebugEvent wiring, C6)
+
+# ── 1. Baseline state ─────────────────────────────────────────────────────────
+/status
+/time
+/npcs
+
+# ── 2. Debug-snapshot builders (the moved modules) ───────────────────────────
+# These commands call build_clock_debug, build_world_debug, build_npc_debug_list
+# from debug_snapshot/build.rs — the primary targets of the extraction.
+/debug clock
+/debug here
+/debug npcs
+
+# ── 3. Movement — pushes a DebugEvent via the moved DebugEvent type ───────────
+go to The Crossroads
+/status
+/debug clock
+/debug here
+
+# ── 4. More movement + NPC state after a tick ─────────────────────────────────
+look
+/npcs
+go to Darcy's Pub
+look
+/npcs
+
+# ── 5. Time advance — exercises the clock and weather debug builders ──────────
+/speed fast
+/wait 60
+/time
+/debug clock
+
+# ── 6. Final round-trip: status + NPC + events snapshot ──────────────────────
+/status
+/debug npcs
+/debug events
+/map
